Use multiline method chaining and assert_equals in tests
Multilining everything makes it easier to navigate and read, and it
avoids having arguments far off to the side after long strings.

Backslash / newline / period format for multiline chaining is used
because it's the method that communicates the most about the fact that
the chain is multilined. With this approach, it's immediately obvious
that it's a chain no matter what line you look at.