Split unicode tests to one each of latex and html
Since both conditions need testing, add a different test for each one.

As of this commit, these two tests both fail since unicode input causes
empty strings to be returned. Also, the output statements in the python
executables are different and if this is deliberate, then a note needs
to be added.